refactor: category api

+++ src/requests/query.filter.request.ts
@@ -1,8 +1,8 @@
 import { query } from 'express-validator'
 
 export const QUERY_FILTER_REQUEST = [
-  query('page').optional().isInt({ min: 1 }),
-  query('limit').optional().isInt({ min: 1 }),
+  query('page').optional().isInt({ min: 1 }).toInt(),
+  query('limit').optional().isInt({ min: 1 }).toInt(),
   query('search').optional().isString().isLength({ max: 100 }),
   query('sort')
     .optional()
